+++ moderngl/old/TextureCube.cpp
@@ -136,6 +136,8 @@ PyObject * MGLContext_texture_cube(MGLContext * self, PyObject * args) {
 	texture->max_level = 0;
 	texture->anisotropy = 1.0;
 
+	texture->reference = nullptr;
+
 	Py_INCREF(self);
 	texture->context = self;
 
@@ -434,6 +436,26 @@ PyObject * MGLTextureCube_use(MGLTextureCube * self, PyObject * args) {
 	Py_RETURN_NONE;
 }
 
+PyObject * MGLTextureCube_transfer(MGLTextureCube * self, PyObject * args) {
+	MGLContext* context;
+	bool args_ok = PyArg_ParseTuple(
+			args,
+			"O!",
+			&MGLContext_Type,
+			&context);
+	if (!args_ok) return 0;
+	MGLTextureCube* new_texture = new MGLTextureCube(*self);
+	Py_INCREF(context);
+	Py_INCREF(new_texture);
+	Py_INCREF(self);
+	new_texture->context = context;
+	new_texture->reference = self;
+	PyObject* result = PyTuple_New(2);
+	PyTuple_SET_ITEM(result, 0, (PyObject*)new_texture);
+	PyTuple_SET_ITEM(result, 1, PyLong_FromLong(new_texture->texture_obj));
+	return result;
+}
+
 PyObject * MGLTextureCube_release(MGLTextureCube * self) {
 	MGLTextureCube_Invalidate(self);
 	Py_RETURN_NONE;
@@ -445,6 +467,7 @@ PyMethodDef MGLTextureCube_tp_methods[] = {
 //	{"build_mipmaps", (PyCFunction)MGLTextureCube_build_mipmaps, METH_VARARGS, 0},
 	{"read", (PyCFunction)MGLTextureCube_read, METH_VARARGS, 0},
 	{"read_into", (PyCFunction)MGLTextureCube_read_into, METH_VARARGS, 0},
+	{"transfer", (PyCFunction)MGLTextureCube_transfer, METH_VARARGS, 0},
 	{"release", (PyCFunction)MGLTextureCube_release, METH_NOARGS, 0},
 	{0},
 };
@@ -618,8 +641,12 @@ void MGLTextureCube_Invalidate(MGLTextureCube * texture) {
 
 	// TODO: decref
 
-	const GLMethods & gl = texture->context->gl;
-	gl.DeleteTextures(1, (GLuint *)&texture->texture_obj);
+	if (texture->reference) {
+		Py_DECREF(texture->reference);
+	} else {
+		const GLMethods & gl = texture->context->gl;
+		gl.DeleteTextures(1, (GLuint *)&texture->texture_obj);
+	}
 
 	Py_TYPE(texture) = &MGLInvalidObject_Type;
 	Py_DECREF(texture);
